ud_project/koohrang-commentrisk/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
backoff==2.2.1
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
et_xmlfile==2.0.0
fastapi==0.115.5
filelock==3.16.1
greenlet==3.1.1
groq==0.12.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
identify==2.6.3
idna==3.10
jiter==0.7.1
joblib==1.4.2
langfuse==2.57.5
mysql-connector-python==9.1.0
nodeenv==1.9.1
numpy==2.1.3
openai==1.54.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
platformdirs==4.3.6
pre_commit==4.0.1
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.10.1
PyMySQL==1.1.1
python-abc==0.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
schedule==1.2.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
textwrap3==0.9.2
tiktoken==0.8.0
tqdm==4.67.0
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
virtualenv==20.28.0
wrapt==1.17.0
